finished (?) split of Xtals code
