+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PorousMaterials"
 uuid = "68953c7c-a3c7-538e-83d3-73516288599e"
 authors = ["SimonEnsemble <cory.simon@oregonstate.edu>"]
-version = "0.3.1"
+version = "0.4.0"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
